Simplify TestDetermineDifferingBitIndex
