+++ pkg/kademlia/routing_helpers_test.go
@@ -696,96 +696,6 @@ func TestDetermineLeafDepth(t *testing.T) {
 	}
 }
 
-func TestDetermineDifferingBitIndex(t *testing.T) {
-	filledID := func(a byte) bucketID {
-		id := firstBucketID
-		id[0] = a
-		return id
-	}
-
-	rt, cleanup := createRoutingTable(t, storj.NodeID{})
-	defer cleanup()
-
-	cases := []struct {
-		testID   string
-		bucketID bucketID
-		key      bucketID
-		expected int
-		err      *errs.Class
-	}{
-		{testID: "A",
-			bucketID: filledID(191),
-			key:      filledID(255),
-			expected: 1,
-		},
-		{testID: "B",
-			bucketID: filledID(255),
-			key:      filledID(191),
-			expected: 1,
-		},
-		{testID: "C",
-			bucketID: filledID(95),
-			key:      filledID(127),
-			expected: 2,
-		},
-		{testID: "D",
-			bucketID: filledID(95),
-			key:      filledID(79),
-			expected: 3,
-		},
-		{testID: "E",
-			bucketID: filledID(95),
-			key:      filledID(63),
-			expected: 2,
-		},
-		{testID: "F",
-			bucketID: filledID(95),
-			key:      filledID(79),
-			expected: 3,
-		},
-		{testID: "G",
-			bucketID: filledID(255),
-			key:      filledID(255),
-			expected: -2,
-			err:      &RoutingErr,
-		},
-		{testID: "H",
-			bucketID: filledID(255),
-			key:      bucketID{0, 0},
-			expected: -1,
-		},
-		{testID: "I",
-			bucketID: filledID(127),
-			key:      bucketID{0, 0},
-			expected: 0,
-		},
-		{testID: "J",
-			bucketID: filledID(63),
-			key:      bucketID{0, 0},
-			expected: 1,
-		},
-		{testID: "K",
-			bucketID: filledID(31),
-			key:      bucketID{0, 0},
-			expected: 2,
-		},
-		{testID: "L",
-			bucketID: filledID(95),
-			key:      filledID(63),
-			expected: 2,
-		},
-	}
-
-	for _, c := range cases {
-		t.Logf("%b\n")
-		t.Run(c.testID, func(t *testing.T) {
-			diff, err := rt.determineDifferingBitIndex(c.bucketID, c.key)
-			assertErrClass(t, c.err, err)
-			assert.Equal(t, c.expected, diff)
-		})
-	}
-}
-
 func TestSplitBucket(t *testing.T) {
 	rt, cleanup := createRoutingTable(t, storj.NodeID{})
 	defer cleanup()
